requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
awscli==1.27.105
Babel==2.12.1
bcrypt==4.0.1
blinker==1.6
botocore==1.29.105
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cloud-init==22.2
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==15.1.0
cryptography==40.0.1
dbus-python==1.2.18
decorator==5.1.1
discord-py-interactions==4.4.0
distro==1.8.0
distro-info==1.0
docutils==0.19
ec2-hibinit-agent==1.0.0
frozenlist==1.3.3
hibagent==1.0.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.1.0
incremental==22.10.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
keyring==23.13.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
more-itertools==9.1.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
orjson==3.8.9
packaging==23.0
pexpect==4.8.0
Pillow==9.5.0
pip-review==1.3.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.23.0
pycparser==2.21
Pygments==2.14.0
PyGObject==3.44.1
PyHamcrest==2.0.4
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
pyserial==3.5
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian==0.1.49
pytz==2023.3
PyYAML==6.0
requests==2.28.2
roman==4.0
rsa==4.9
s3transfer==0.6.0
SecretStorage==3.3.3
service-identity==21.1.0
six==1.16.0
sos==4.3
ssh-import-id==5.11
systemd-python==234
Twisted==22.10.0
typing_extensions==4.5.0
ubuntu-advantage-tools==27.10
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.15
validators==0.20.0
wadllib==1.3.6
yarl==1.8.2
zipp==3.15.0
zope.interface==6.0
